LBV2: Don't cascade Pool deletes to VIPs
Deletion of a Pool via the LBV2 API was incorrectly cascading to the VIPs
associated with the Pool. This seems to have been the intended behavior from V1,
so rather than changing the underlying binding, which would affect LBV1 as well,
I'm just adding an update operation to clear the pool's VIP IDs before deleting
it in LBV2. Since the vipIds list is empty, the delete won't cascade.

Signed-off-by: Brandon Berg <bberg@midokura.com>
Fixes: MI-1950
Change-Id: Ic98c5d91d966d7dcc5445f2bc7dca4226233f72f