Use container commmons exec to run vppctl
Using ProcessHelper is exepensive for short tasks as it will
create two threads in addition to the underlying thread created
by UNIXProcess, and these are used only to drain the process
output streams. This is done while the calling thread is waiting
for the process to exit.

The ContainerCommons trait provides an alternative lightweight
method to launch an external process, where draining the output
streams is performed on the calling thread.

This changes the VppCtl class to a trait that leverages the
ContainerCommons to execute vppctl commands.

Change-Id: Id2d7e689c376ff33653f756fa608542ef5122fb2
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>