State proxy server requires specific address
This changes the state proxy server such that it requires a
specific IP address configuration before binding to a local
address-port pair and publishing it in the service discovery.
To this end, this patch introduces a new configuration option
that allows the operator to specify a local interface name
rather than an IP address that will be used to bind the state
proxy server.

The new configuration works as follows.

* If the address is configured with any value other than any
  local IP address, the state proxy server will use that
  address.

* Otherwise, if the interface is configured and it has only
  one IP address, the state proxy server will use that
  address.

* Otherwise, if the interface is configured but it has none or
  multiple IP addresses, the state proxy server is disabled.

* Otherwise, if neither the address nor the interface are
  configured, the state proxy will start only if the host has
  a single unicast IP address.

* Otherwise, the state proxy server will be disabled.

Change-Id: I23ef32c831835177d0f42c816b66cade4c3d9bf7
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>