Fix race to clearing the admin token
When the admin token is obtained with password credentials,
if an unauthorized exception occurs with such a token during
a Keystone request, the client will clear and renew the token
once in order to prevent being deadlocked with an expired token.

However, multiple concurrent requests may race when renewing
the token, invalidating an assumption that the previous token
is always non-null, which may generate NPEs. This fixes this
check by using a local copy of the admin token that cannot
be modified by a concurrent request.

Fixes: ME-1365

Change-Id: Ie7028809ced24c099083c2cb27c1fdeda774b83b
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>