Add a read timeout to the state proxy client
If the state proxy client doesn't hear from the state proxy server it is
connected to within the configured read timeout period (default 10s), it
disconnects.

To distinguish failure cases from cases where the server is simply
quiet, the client send a ping to the server at an interval of one third
the read timeout period, which a well functioning server should respond
to immediately.

Change-Id: Ied2fc7a951a44e647cadccaa856694500764b33c
Signed-off-by: Ivan Kelly <ivan@midokura.com>