Add link option to set master
This patch modifies the link utility classes so that we can add
a master option. This enables us to bind an interface to a given
linux bridge.

Doing a setMaster(link, "br0") would be the equivalent of binding
that link's interface to the bridge with name "br0".

Or using a linux shell:
ip link set <if-name> master br0

Or:
brctl addif br0 <if-name>

Ref: IIOT-255

Change-Id: I08c5db2c58eeea048c12c4fc6066bdd0b450bddb
Signed-off-by: Mateo Hermida <mateo@midokura.com>