Update patch set 1
Patch Set 1:

Thanks, Alex. The use case of this is to allow agent minions created outside of midonet core (for instance, by MEM or IIOT) to have any arbitrary kind of parameters. Right now, this is impossible to do without changing the bindings in AgentServiceModules.scala. With this patch, bindings can be added to the injector used to provide parameters to agent minions without any further changes to midonet core.

The concrete use case we have right now is that we're implementing in the IIOT team an agent minion that has a few parameters whose module bindings are not added in the current midonet code. If there's already a way to add the class->instance bindings without changing midonet, please let me know. Otherwise, that's your use case :)

Patch-set: 1
Label: Code-Review=0