ZOOM writes data provenance
With this change, ZOOM adds provenance information for changes
to ZOOM objects. The provenance is a historical list of ownership
for the changes to the objects and includes the product version,
GIT commit SHA, change owner and change type (data or backreference).

New provenance entries are always appended to the current list,
unless the provenance for a change is the same as the last one.
A provenance is the same if they have the same change owner.

If the new provenance is the same (same owner) but either the
product version, product commit SHA, or change type have changed
then the last provenance entry is updated with this new
information.

Ref: MI-1989
Ref: MI-1990
Ref: MI-1991

Change-Id: Iaff0ef244034a2c044d6861866f5216bcf99fc0d
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>