Create Vpn SeriveContainer on first siteconnection
If VpnService container is created without site connections
it state changes from "STARTING" to "STOPPED", so container scheduler
kills it within 10 seconds and tries to reschedule same container on
next host within 15 seconds. This is done in a loop until first site
connection is created.

This patch avoid this resource wasting by creating
vpn service container when first connection is created and deleting
it when last connection of all vpn services that use this container
is deleted. The containerId field of VpnService objects is managed by
SiteConnection translator.

Ref: MI-1230
Change-Id: I42b30b106bb392279f1f00c1dec4ccc333244b02
Signed-off-by: Dmitry Bufistov <dmitry@midokura.com>