Introduce CacheableStorage
This patch introduces the CacheableStorage trait,
which is inherited by InMemoryStorage and by the
ZookeeperObjectMapper. Similarly to Storage.observable,
the CacheableStorage.cache(c: Class[T <: Obj]) allows
you to receive notifications when objects of a given
class are created, updated, or deleted. Differently
though, it provides a flat Observable whose emitted
events are of the type Map[ObjId, T], with an up-to-date
copy of all objects of class T.

Change-Id: Ic99caa9604673393db87c0a9f7f1fa6176bd6031
Signed-off-by: Eduardo Bezerra <eduardo@midokura.com>