Update patch set 3
Patch Set 3:

Wrt to the C3POStorageManager, we don't support/maintain it at the moment (other than usage in some unit tests), so I left it as is.

I'll add the ownership in the next patch set to make sure this something we don't overlook in the future, but I guess this depends on whether we'll ever use the Neutron importer as intended.

Patch-set: 3
Label: Code-Review=0