Container service uses RX backpressure
This fixes an issue in the container service, where a backpressure
buffer was not used for scheduling RX notifications on the
container service thread. Since this thread also performs I/O
operations on behalf of the container delegates, it may be blocked
for enough time such that the 128 entry buffer provided by the
observeOn operator to fill. While an additional backpressure
buffer was used with observeOn in several cases, this was not
so in all circumstances. This patch addresses this omission.

Change-Id: I12261c4ea8a8cea74100f51456b3cddf168839dd
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>