VPNaaS: Report container status on zero config
This fixes an issue in the IPSec container, where if the container
does not launch libreswan because the Neutron VPNaaS configuration
is incomplete, the container reports an ERROR container status
back to the cluster, or doesn't report any status at all. In turn,
this determines the cluster to keep on rescheduling the container,
even if subsequent attempts lead to the same result.

This patch changes the container logic such that, when the
container configuration is incomplete, the container nevertheless
reports a RUNNING status telling the cluster that the scheduling
has been successful.

Partially-Fixes: MI-2022

Change-Id: I6b197e541788c17baadf1dc1faf52c5103ec5f00
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
(cherry picked from commit 40b7ba02e62251e76076495b374a0d710af4bf14)