Prevent device manager name collisions
A race-condition exists when a virtual topology device manager is
started and stopped in a quick succession. If the manager actor has not
stopped by the time it is started again, an InvalidActorNameException is
thrown, causing a restart of the agent.

This patch fixes the issue by ensuring that every manager actor is given
an unique name.

Ref: MI-1998
Change-Id: I9fb6b05c48c5d175ed7c3942d580d7bac3f47d80
Signed-off-by: Adrian Serrano <adrian@midokura.com>