Flow state client should always close socket
Previously the client was using ClosingRetriable, which was closing a
socket member variable. However, if a retry did happen, it would
overwrite this socket variable with a new instance before the close
happened, so sockets were leaked.

This change removes ClosingRetriable from the flow state client, and
instead replaces it will explicit calls to close from a finally clause.

Ref MI-1905

Change-Id: Icf586e1102dd8ca341f41ef2420ffa6cf33dbdf2
Signed-off-by: Ivan Kelly <ivan@midokura.com>