Throttle flow state data reads at minion
If the flow state minion writes flow state data to the client faster
than the client can consume it, the datas will build up in memory and
eventually trigger and OOM error.

This change throttles the data so only there can only be 1MB of data
outstanding before we ensure that the client is consuming the data.

Ref MI-1905

Change-Id: Ie90d9863674d14b460bd1a7327b1cfa56e52f634
Signed-off-by: Ivan Kelly <ivan@midokura.com>