Fix order of jump rules on post-routing chain
The fix for MI-1979 (commit 932c7bae) mixed up the order of the jump rules on
the post-routing chain, moving the jump to the skip-SNAT chain to the end, which
caused SNAT not to be skipped when appropriate. Fixing this issue and enhancing
an existing unit test to verify order.

Change-Id: I2385261e6846254b05fa2040ed0c9236fec7691d
Signed-off-by: Brandon Berg <bberg@midokura.com>
Fixes: MI-2024