Netlink: Fix parsing of kind attribute
Netlink assigns a 7 byte array to the kind attribute of a link. If the
actual kind is smaller, for example "veth", then it assigns an extra 3
zero bytes in the buffer for that attribute, resulting in the string
"veth???". This means that making a comparison between the constants
defined in NestedAttrValue.LinkInfo and this value will fail most of the
times.

This patch fixes the creation of the string representing the attribute
by not using these zeroes.

A test is updated to test for this condition too.

Change-Id: Ia15e8052535b6b6f8427b129493ab48b0c6bab6b
Signed-off-by: Mateo Hermida <mateo@midokura.com>