Allow pluggable insights data collection
This change allows the MidoNet agent to load a pluggable
insights component to report flows, simulations and local
metrics.

To this end, the patch defines a package local interface, that
custom analytics components may implement, in order to get access
to the agent's metric registry, and receive calls for completed
simulations and flows. The implementation should provide an
inexpensive, thead-safe way to fetch this data and move it
off process for further consumption.

Change-Id: I0407b8da8032b9b9ab5d66636da886dd1a1b2a47
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>