Add MDTS openstack provisioning script
This script provision a neutron openstack topology
in mdts. It configures a external network with range
200.0.0.0/8 and an edge router that peers with the
quagga containers using BGP. Midolman containers are
configured in the same tunnel zone and can be used as
computes, using namespaces to emulate vms and attaching
them to neutron ports with the command mm-ctl. You can
access this fake VMs from the MDTS quagga containers,
assigning Floating IPs in neutron.

Also it adds a ligther flavor without the L2GW, vtep
and cassandra containers to be used for openstack
testing only.

Change-Id: I6b7281b20d5ba5280b54560b3450cbdf6f7e8b4e
Signed-off-by: Antonio Ojea <aojea@midokura.com>