Transactions: allow cascade delete with no backreferences
Right now, we don't support cascade deleting on a transaction
if the right side of the binding has a null in the fieldName,
which may be usefull to support only forward references (where
adding backreferences is not possible).

This patch removes that restriction taking care of not clearing
the backreference if there is none specified in the binding.

Change-Id: I386c49061aec2265b260cb6d1e03cf00d2697ad4
Signed-off-by: Xavi León <xavi.leon@midokura.com>