Fix getting port address from JSON
The fromProto method is only called when bringing in data from
ZK and converting it to a POJO.  It will not get called when
data comes in from the REST API.  That will be deserialized JSON
data from Guice, and will not run fromProto, or, more importantly,
afterFromProto.  This means that the portAddress of a port wasn't
getting set when the REST API was sending data, although when
pulling from ZK, the port address was getting set correctly (as
afterFromProto was getting called).

This led to the port route which gets created in the
PortResource's "createFilter" method was getting passed in a
blank port address every time, resulting in an invalid port route
on the router (no dst network addr was getting set).

Fix is to make a common function and call it both when the Port's
"create" method is called (typically from the REST API when first
creating the object in a POST) and in the afterFromProto function
when data comes in from ZK.  This way, we will set the port
address effectively when data comes in from both the REST API and
from ZK.

Change-Id: I04034cbe30468b3b3f53768e0f304a60dda326e5
Signed-off-by: Michael Micucci <micucci@midokura.com>