Fix possible NPE in NovaMetadataClient
The response.getType() call can actually return NULL if the
'Content-Type' header isn't defined.  To prevent the NPE, we
can check if that is null and if it is, return some default (like
UTF-8), otherwise return the charset parameter as normal.

The other usage of "getType" is just to print a log message, so
it should properly show up as "null" there.

Change-Id: If3dfc9af2d5d87eb00a95f4982a3b3240052e696
Signed-off-by: Michael Micucci <micucci@midokura.com>