Change state proxy warning to only print once
The State Proxy would print a warning every time it tried
to reconnect and failed (which would happen every second until
failure unless the proxy server was set up).  First, as this
was a StateProxy-specific log message, the log.warn was moved
from PersistentConnection to StateProxy itself. Second, the log
now only prints on the first miss (when retry count is 0)
instead of each iteration, and it iwll print out with a count
of how many iterations it will attempt.

Change-Id: Ifce4e905654e6af5c53a40c837f5ddbfe4637a4c
Signed-off-by: Michael Micucci <micucci@midokura.com>