Zoom: Use getIdString instead of directly the ObjId
In Zoom, we use this getIdString to get a UUID string regardless
of the actual ObjId passed by parameter (either a String or a
Protobuf that gets converted to a UUID String). This patch fixes
the only place where we were specifically assuming a UUID string
without type checking, and uses the getIdString to make it coherent
with the rest of methods.

Backported from master

Ref: MINOR_CHANGE

Change-Id: Iade434f797f4420f3355560cf109663633543bfe
Signed-off-by: Xavi León <xavi.leon@midokura.com>