VPNaaS routes use network address rather than subnet
This bug crept in during the backport of the fix for
MI-1230 (87037b). The new translation code was calling
routerPortAddress, passing in a IPv4Subnet. However, this subnet was
usually the port address and not the network address, so the returned
IPv4Addr was incorrect, as it expects to be calculated from the network
address. For example, if 169.254.0.1/30 was passed in, 169.254.0.3 would
be returned, which is actually the broadcast address for that network.

This issue did show up in a unit test during the backporting, but just
changing the test fixed it, so I mistakenly assumed it a non-issue.

Ref MI-2152

Change-Id: If5b62c5d833234941879463d2ab25dd7ac7c864e
Signed-off-by: Ivan Kelly <ivan@midokura.com>