Enable ipv6 in on mdts containers
A docker updated started automatically disabling ipv6 which interfered
with the pipework created networks we use for bgp and ipv6 testing.

First attempt at a fix was to switch these interfaces to using docker
networks. However, adding docker networks doesn't allow you to specify
the interface name, which is important for our usecase. Even more
importantly, docker networks seem to apply some sort of ip antispoofing,
so the interfaces created cannot be used for routing, so are thus
completely unsuitable for our testing.

The solution herein is simply to enable ipv6 on all interfaces in
containers which forward traffic.

Change-Id: I1ac16966ba6446e3845c25b0fe5897dcd8b287ea
Signed-off-by: Ivan Kelly <ivan@midokura.com>