Don't use cached MAC for VLAN-tagged packets
For packets which are VLAN-tagged, do not use cached results on
a bridge to reply to an ARP request.  This is because the generated
reply will NOT be VLAN-tagged, and will result in an incorrect
ARP reply packet which cannot be consumed by the ARP-request
sender.  Instead, by forcing a cache-miss, the ARP reqeust will
get braodcasted as-is to the bridge's ports, allowing the actual
destination with that IP to reply with an ARP reply packet, which
will be correctly formed in the case of VLAN-tagging.

This also goes for any ARP request from a VLAN to a router port,
as there should never be a case where a normal bridge will need
to send a VLAN-tagged packet to a router (as there would be no
way the router would be on the same subnet as the VLAN in this
case), and a VLAN-aware bridge would still send the proper ARP
request to the correct router port in a multicast, anyway.

This should have no side-effects when dealing with VLAN-aware
bridges as they should not have any direct ports, other than
the router port mentioned above, which will be dealing with
VLAN-tagged packets anyway.  All VLAN-tagged packets into a
VLAN-aware bridge will be sent out a L2-connection to another
bridge or router, and thus need no MACs cached for any endpoints.

This should also have no side-effects for normal packets, which
will still try to use cached information.

(Backported from master)

Ref: MI-2173

Change-Id: I125796507523d7a18007c36089ee695547407851
Signed-off-by: Michael Micucci <micucci@midokura.com>