Binary flow recorder decodes messages with no IP
Flow recording can record messages where there is no L3 info, such as
LLDP messages. This should be handled gracefully. Previously we only
handled IPv4 and IPv6 messages, so if a message was found without any L3
info, an error would be thrown on decode.

This patch adds a "Null" InetAddrType which is used in the case that
there is no L3 info.

(Backported from master)

Ref: MI-2020

Change-Id: I14ff6536d848a8858c788570f31806b15f350a2b
Signed-off-by: Ivan Kelly <ivan@midokura.com>