Re-structure helper func to allow for tailrec
The "force" helper test function wasn't doing tail recursion as
the recursive call to "force" was inside a catch (which is
not a tailrec position).  Instead, move the recursive call out of
the exception handler and add the @tailrec annotation to ensure
this function is tail recursive and won't blow the stack if a
long timeout forces many loops of this function.

Change-Id: I9581e102f1ede2ef4f9aad9c690c0c93ff728b58
Signed-off-by: Michael Micucci <micucci@midokura.com>