RoutingManager to use separate BGP connection
If midolman loses its connection to Zookeeper, it will attempt to
re-establish the connection immediately. The BGP management
infrastructure uses the same thread as the simulation to monitor the
connection and process watchers. The BGP management infrastructure will
clean up its BGP namespaces if it cannot connect in a couple of seconds.

The problem here is that the BGP management infrastracture will assume
that the reconnect has failed if it's watcher has not been processed by
the time that the asynchronous timer has fired. If there is a large
topology, a reconnect to zookeeper may have actually happened, but the
large amount of watchers to process will keep the BGP infrastructure
from getting it's watchers processed.

The solution is to give the BGP management infrastructure it's own
thread and connection to Zookeeper.

(Backported from master)

Ref: MI-2071

Change-Id: Ieb31e6a00778353497b96242558e6d33497f4127
Signed-off-by: Joe Mills <joe@midokura.com>