Add max_attempts as parameters to setup_package
This patch changes the __init.py__ file of the mdts
module so that a max_attempts parameter can be given
to the setup methods. By doing this, other modules
that reuse the setup of mdts can configure a higher
(or lower) wait for the mdts environment to be ready.

(backported from master)

Ref: MINOR_CHANGE

Change-Id: I67cc16d5aed50b805d6c4dea96e000167444d2c7
Signed-off-by: Eduardo Bezerra <eduardo@midokura.com>