mm-ctl fallback to legacy when minion unavailable
When we went from having the mm-ctl connect directly to zookeeper to
having mm-ctl connect to a local minion which then connected to
zookeeper. This was down to avoid the startup time involved in bringing
up a JVM instance.

However, this didn't handle the case where the minion itself was
unavailable, such as when the agent was stopped, or the minion had
crashed.

This patch adds a fallback for that case, where, if binding though the
minion fails, we fall back to the legacy method which connects directly
to zookeeper.

(Backported from master)

Ref: MI-2179

Change-Id: Ic8649e7904aa709ef993e47c7401a7e15b399fbc