Thread safety for native timed expiration map
Locking is very coarse right now. Each operation locks the map. This
shouldn't be more expensive than the java implementation as we use a
ConcurrentHashMap there, which has plenty of locking. However, it may
mean more contention when obliteration runs. If this proves to be a
problem, we can solve it then by locking the expiration queues and ref
map separately.

Ref: MI-2225
Change-Id: Ibbb353e3b1979dfc66bbb5460abdd17a4ce1641a
Signed-off-by: Ivan Kelly <ivan@midokura.com>