Remove FlowIndexer interface
This interface did nothing but make the flow controller more complex
with abstract override junk.

The flow tag and expiration indexers have now be moved to members of the
flow controller.

Calling expiration on the indexers now returns the next flow to
expire, rather than removing it directly itself. The caller should
ensure this has been removed from the flow table. Expiration should be
called repeatedly until it returns null.

Similarly, flow tag invalidation returns a list of flows to remove,
rather than removing them itself. It does however remove the tag-to-flow
associations for the invalided flows from the tag-to-flow table. The
caller of tag invalidation is responsible for removing the flow from the
flow table.

Change-Id: I3f179ce6b0d8b92a80aecb4f5db89529fe0a1b8c
Ref: MI-2265
Signed-off-by: Ivan Kelly <ivan@midokura.com>