Support for zebra v3 protocol
The zebra v3 protocol adds a new VRF field in
its headers, which makes it incompatible wth the
previous version of the protocol.

This change adds support for either version of the
zebra protocol, sending the default value for VRFs
since they are currently not used in midonet.

The interaction with some previous communications of
the protocol also changes. For example some new messages
are added, and to subscribe to the route_add messages,
we need to register the zebra server as a next hop.

Ref: MI-2284
Ref: MI-2342

Change-Id: I038e5672b245fdaf2a9f3c8427d5a5a3e1d1ea10
Signed-off-by: Mateo Hermida <mateo@midokura.com>