FlowTags generate 64bit hashes
FlowTags are used to index flows, but the we never use the flow to look
up the value of the tag. Therefore we don't need to use the full tag for
indexing, we can use a id which is "uniquely" generated for the tag.

If we use a 64bit hash to for the long, and there are will need to be
600,000 tags for the chance of a collision to be 1 in 100million [1].

In the case of a collision, all we will get is some spurious flow
invalidations, which would not be a big problem.

Using hashes will make the flow tag index much more compact in memory,
and simplify sending the flow tags off heap.

[1] http://preshing.com/20110504/hash-collision-probabilities/

Change-Id: I9d51048a85f509f9f54a4465e706edfee568cdae
Ref: MI-2269
Signed-off-by: Ivan Kelly <ivan@midokura.com>