Do not force type annotation on overriden methods
The scalastyle_config.xml forces the definition
of return types for all public methods, including
overrides. While this may make sense for traits
and interfaces, most of the original code does
not follow this for the overridden methods, causing
some editors (particularly IDEA) to complain and
potentially hide other issues. This patch adds
a parameter to the rule to ignore overrides.

Ref: MINOR_CHANGE

Change-Id: I548841123d77f1a03305c97249a299ced50f837a
Signed-off-by: Ernest Artiaga <ernest@midokura.com>