More granular locking for timed expiration map
Benchmarking showed that the interthread access were causing threads to
yield very frequently with a negative impact on flow processing
latency. This is because each time a stateful rule is traversed, all
threads are queried for matching flow state if none is found on the
local thread (which is always the case for the initial packet in a
connection).

This change changes the native timed expiration map to remove the lock
and use a concurrent hash map from tbb instead. The implementation now
looks very like the scala version that uses ConcurrentHashMap.

Change-Id: If4c866743b625009eb875d95cd1c165eb3efa913
Ref: MI-2263
Signed-off-by: Ivan Kelly <ivan@midokura.com>