Only submit flow state to a single worker
Until now flow state has been sent to all workers. This is dumb. If a
worker is looking up flow state, and it doesn't find it in its local
table, it looks at the other worker's tables anyhow.

The only possible reason I can think of for sending to all threads is to
try to keep this lookup in the cache. But these are large hash tables,
with a random memory access pattern. The keys you look up likely aren't
in the l*-cache, so reading from another workers table is likely to be
as quick as reading from your own table, NUMA notwithstanding.

In any case, the cost of processing a piece of flow state at all workers
is much more than the cost of jumping into another workers table, since
most flow state is never used after the first use.

Change-Id: I2c6cc52f0e766f80fb67cc69f93abe30706c17f6
Ref: MI-1273
Signed-off-by: Ivan Kelly <ivan@midokura.com>