TBB as shared library
The metering c++ implementation will need to use TBB its metrics map, as
it needs to be written to by one thread (the worker thread), and read
from by another thread (the http server).

Previously TBB was used as a static library, which was compiled into the
timed expiration map library. However, if the metering implementation
needs it, a static library cannot be used, as it would mean multiple
copies of the code in the process, which TBB explicitly warn against.

Change-Id: Ic0bebb45490f7b7b14f87ecc4f95c7154a626fc4
Ref: MI-2313
Signed-off-by: Ivan Kelly <ivan@midokura.com>
Signed-off-by: Xavi León <xavi.leon@midokura.com>