Zebra: Read full message if not implemented
In case of receiving a not supported message, we were leaving the
remaining bytes on the input stream, which made the following messages
to start on the wrong position. This patch reads the number of bytes
declared in the header so the input stream is flushed and leave it ready
for the next message.

Change-Id: I737006d79e64877a88bf1c22c6f22c365cf555fc
Ref: MI-2284
Signed-off-by: Xavi León <xavi.leon@midokura.com>