Add scripts to generate release notes
This patch adds a script that allows the user to generate
a list of changes from two points in the git history, and
print the jira issue description or in case of a minor change
print the commit message.

A valid Jira account is needed for the auth against the Jira
REST API.

Usage:

The change_log.py file needs to be edited and username and
password updated locally.

Afterwards do a:
./change_log.py <new_release> <previous_relase>
Where new and previous release are the commit ids of the
release that we want to make and the previous one. These
can be commit hashes, tags or branches.

Change-Id: I0e4616a440ae54997bd9e65a460154cf8d124429
Ref: MI-2231
Signed-off-by: Mateo Hermida <mateo@midokura.com>