BGP: just consider the port associated to the handler
Right now, we were storing the mapping from routes to
virtual ports because the RouterBgpMapper published all
the bgp ports for a given router. However, we associate
a single RoutingHandler to each bgp port, so there's no
need to store that mapping internally. This patch
simplifies this class by removing the not needed code.

Change-Id: I6f05b37a7f39a255d685157ff1f0b0f745119dfb
Ref: MI-2405
Signed-off-by: Xavi León <xavi.leon@midokura.com>