BGP: remove route even if next hops empty
BGPd sends us an IPV4_ROUTE_DELETE when there's only one
next hop for a given route. If we receive that message, it
means that there's no path for that route. In that case,
we need to remove it from the topology.

Previously, we were checking that the IPV4_ROUTE_DELETE
message contained the list of next hops for the route
being deleted, but this is not the case for the zebra
protocol V3 (used in Quagga 1.1.x), in which case the
list of next hops was empty, and thus we were not deleting
that route.

This patch removes the route regardless if the message
contains next hops or not (valid for zebra protocol
V2 and V3).

Change-Id: I07a9d0f894e31efd7f5b6fd7d3c0a950955fab5f
Ref: MI-2405
Signed-off-by: Xavi León <xavi.leon@midokura.com>