AgentReboot: JNI interface for off-heap flow list
The JNI interface exposes two basic methods. Push to
the list and pop from it. To make it memory friendly,
we do not allocate memory when pushing, just pass the
ByteBuffer to the native implementation, which makes
a raw copy according to the boundaries specified.
Deserialization is done when retrieving elements from
the list, so we can use a FlowMatch object to invalidate
the specific flow.

Change-Id: I3fe4f6dd9a4f25d6a2cf4f8a0bdaa2b65ea959bf
Ref: MI-2383
Signed-off-by: Xavi León <xavi.leon@midokura.com>