AgentReboot: introduce flow iterate API on ovs connection
This patch introduces an interation API for retrieving the
list of flows. This way, we can execute the callback for
each retrieved flow. In the case of the agent reboot, this is
usefull because we can directly copy off heap the flow match
for the given flow without deserializing and allocating on heap.
In following patches, we will provide the callback that copies
offheap the flow match.
Otherwise, with the current enumerate API, we must retrieve the
whole list, allocate it on heap, then copying it offheap for
later retrieval.

Change-Id: I6c592d9f11dffff9303db46f0614e67c9b59abbb
Ref: MI-2054
Signed-off-by: Xavi León <xavi.leon@midokura.com>