Cluster: introduce a topology cache minion
This patch introduces a minion that starts monitoring
the topology objects in Zookeeper and allows getting
a snapshot. The subscription to the caches MUST be
done in the same thread receiving the updates from
Zookeeper to ensure a consistent state (done).
In the same way, processing the snapshot outside of the
Zookeeper thread requires the snapshot to be first
serialized to a different object to prevent concurrent
modifications (to be done.)

Change-Id: I9f5dcb0711aabbad81b045cbf39340100793fcc5
Ref: MI-2529
Signed-off-by: Xavi León <xavi.leon@midokura.com>