Cached storage wrapper
This patch implements a wrapper over the Storage and
the StateStorage interfaces. This wrapper utlizes either
a local cached version of these interfaces, or the regular
ZK implementations, if the cache has already been invalidated.

This allows us to use the cached objects while available,
without making further changes in the logic of the agent.

This patch doesn't use the final version of the cache, just a
placeholder for when the actual communication between the agent
and the cluster is completed.

Ref: MI-2538

Change-Id: Icd122043608f6cdac70997e875c5e88fcc0fb9b3
Signed-off-by: Mateo Hermida <mateo@midokura.com>