Introduce a byte buff handler for endpoint minion
This patch introduces an HTTP byte buffer handler which,
upon receiving a request on its configured endpoint, will
reply with the contents of a ByteBuf provided by the
byte buffer provider. This will be later used jointly
with the topology cache minion to transfer the current
state of the topology as a binary stream to the agent.

Change-Id: I3118eea79e9b3dd50ecd8ad95dd14a411369599e
Ref: MI-2531
Signed-off-by: Xavi León <xavi.leon@midokura.com>