Upgrade SBE library to 1.7.0
... so we can encode variable length data inside groups and not
only at the end of an sbe message. This will come in handy when
implementing the topology snapshot sbe message, which contain
variable length data, e.g. strings representing topology objects.

This change also implies some changes in the current code to
accomodate the changes in the encoding/decoding classes.

Change-Id: I260af66b28781b24061d6c3b1a28c0567064cf61
Ref: MI-2578
Signed-off-by: Xavi León <xavi.leon@midokura.com>