Introduce the topology snapshot sbe message
This SBE message encodes a complete snapshot of the
NSDB topology, containing a group of topology objects
and a group of topology state objects. This can be
extended later with topology configuration data if
necessary. This message will be used to exchange
cached topology data between the Midonet Cluster and
Midonet Agents.

Change-Id: I08f82b545423d52a84e252cf1e0716152b17da4f
Ref: MI-2578
Signed-off-by: Xavi León <xavi.leon@midokura.com>