Remove child injector in Midolman
Due to some dependencies across modules, the
MidolmanModule required a first injector to be
set up, and then its contents had to be copied
into a definitive child injector.

This patch makes the necessary values available
at an earlier point (by taking into account how
the module is actually used), removing the need
for an intermediate injector and contributing to
reduce the agent start-up time.

Ref: MI-2547

Change-Id: Ic32dd1431e6c5f8a4d91e017ba97650f8b420861
Signed-off-by: Ernest Artiaga <ernest@midokura.com>