Remove reflection scan during midolman startup
This patch removes the need to perform a (costly)
scan of the classpath during the start-up of the
midolman agent.

Reflection was used to locate container classes used
by the VirtualToPhysicalMapper. Since these classes
are fixed, the reflection scan has been replaced by
a Reflections object explicitly limited to the
relevant classes.

Ref: MI-2540

Change-Id: I299a159a228242bca20935cb4d0548baf7a3e9f0
Signed-off-by: Ernest Artiaga <ernest@midokura.com>