Introduce topology snapshot serializer
This patch introduces the topology snapshot
serializer and deserializer. The SBE shcema
is modified to accomodate the deserialized
object to what the agent may expect, so it
can be directly used as a cache without
further transformations.

NOTE: the TopologySnapshot serializes a
Zookeeper node raw data (ChildData) whereas
it is deserialized as a Protobuf topology
object.

Change-Id: I4e5732ab601ee3886e9ce0ae45b25b9b00282273
Ref: MI-2578
Signed-off-by: Xavi León <xavi.leon@midokura.com>