Introduce a topology mapped snapshot
Until now, a topology snapshot was a simple
list of objects. However, to simplify the
deserialization of the topology snapshot on
the agent side, we introduce in this patch
a mapped snapshot following the cached
topology used by the agent. With this kind
of snapshot, no further transformation will
be required once the SBE message is
deserialized.

Change-Id: I01ef52194de4af2ad4ca69a7e9103f0dbae0061a
Ref: MI-2578
Signed-off-by: Xavi León <xavi.leon@midokura.com>