Stop routing manager actor correctly
The routing manager actor has several child actors that
are created for each new BGP port on midonet. These
children actors however were never stopped so when
the actor system is stopped it always timed out if
there was any of these ports still remaining, making
the shutdown slower.

This patch fixes this by adding a stop to all these
actors once the parent is being stopped. This stop
properly handles the closing of the ZK connection and
the stopping of the child actor.

Ref: MI-2602

Change-Id: I8b775c765e568365257ce4f066125bd1286cf3e3
Signed-off-by: Mateo Hermida <mateo@midokura.com>