Clean-up stale datapath ports
When the existing midonet is 'reclaimed' by the
Midonet agent instead of being re-created, there
may be stale ports from previous executions
corresponding to VMs that do not exist anymore
(and notifications were not received because the
agent was down).

This patch introduces a call to clean up the
stale ports at the end of the DatapathController
initialization.

Ref: MI-2387

Change-Id: I37e8c3de7d0a8966b71e90aaea461b31d0e6f99d
Signed-off-by: Ernest Artiaga <ernest@midokura.com>