Add support for compression in HTTP handler
This patch adds a handler in the HTTP netty channel pipeline
that supports the compression of the data sent. This will enable
the compression of the topology snapshot, since the data will be
compressed when sent via this handler, and the topology client
already supports data decompression when a request with the
appropiate headers is received.

Ref: MI-2532

Change-Id: I3d79af1ffb8c71633a2b296ee9dda8d8a53d8430
Signed-off-by: Mateo Hermida <mateo@midokura.com>