Remove reflections from container framework
Even though we were specifying specific classes on the
Reflections contructor with the hope of reducing the
scan time, it still needs to scan each file on each
resource (.jar file) and check if the file matches the
package of the classes passed by parameter. This is
still time consuming.

This patch removes reflections from the container
framework given that the amount of container types is
manageable. It certainly reduces the flexibility of
the container framework at the cost of speeding up
the agent boot sequence. If we ever need to add this
flexibility back to the container framework, we should
defenitely move this framework to the Midonet Agent
companion process (a.k.a agent minion).

Change-Id: I64245ac0c59b2758e60039183de886ab815ce163
Ref: MI-2550
Signed-off-by: Xavi León <xavi.leon@midokura.com>