Fix progress issue in topology snapshot provider
There was a missing case which prevented the topology
snapshot provider to make any progress. Specifically,
the snapshot provider was waiting for a future to be
completed by a client request. If there was no client
requests, no one would complete the future and left
the snapshot provider state as in-progress indefinitely.
Subsequent requests would be enqueued waiting for the
snapshot encoding to finish, which would never happen.

This patch handles the cases where:
i) there's a snapshot in progress -- skip encoding.
ii) there are client requests in progress -- wait until
    all of them finish.
iii) there is no snapshot in progress and no requests
     in progress -- proceed with the encoding.

Change-Id: I76e098d6c47ad0123394640fad0373d3451ed57c
Ref: MI-2614
Signed-off-by: Xavi León <xavi.leon@midokura.com>