Make virtual topology an abstract service
During the instantiation of the VirtualTopology object, we
were accessing the discovery service which requires an active
ZK connection. However, the MidonetBackend holding such
connection is not started until later.

This patch changes the VirtualTopology to be an abstract service
and moves the initialization of the topology snapshot, storage
and state storage when we start the service itself (after the
MidonetBackend service is started).

This patch also makes fetching the topology snapshot from the
cluster a sequential operation instead of an asynchronous
operation. This avoids context switching and buffer copying
between threads, which can take some time on large topologies.

Change-Id: I6ace592b6f0e6f341247a9c806ddb34835278968
Ref: MI-2615
Signed-off-by: Xavi León <xavi.leon@midokura.com>