Call ref() on ManagedFlows from the oversubscription pool
We were not calling ref() on the ManagedFlows taken from the
oversubscription pool. When returning the managed flow to the
pool, calling unref() threw an exception complaining about that.
Just calling ref() when taking the managed flow from the
oversubscription pool fixes the issue.

This code is soon to be dead-code as we use off-heap data structures,
but it's not still deprecated.

Change-Id: If10e2c5db7bf294562a29a50cd239fa6828a2d4e
Ref: MI-2618
Signed-off-by: Xavi León <xavi.leon@midokura.com>