Remove compression from topology snapshot client
In a very big topology, the GZIP compression turns it from
a 20MB request to about 5MB. However, the compression done
makes the request last around 1.5 seconds instead of a couple
of milliseconds. It's not worth the extra boot time in order
to save sending the extra MB for a one-time-only operation.

Ref: MI-2007

Change-Id: I2d50d8dbb54c28251358d8f4a4d630a6dc567d08
Signed-off-by: Mateo Hermida <mateo@midokura.com>