Allow empty values on state snapshot
We only listen on changes for Host and Port state objects.
However, the agent may query other state type objects
(e.g. PoolMember) for which we don't cache the values.
Ideally, we shoud change how we listen on changes to the
state objects as currently there's no way to cache the
owner of a state key for Pool Members -- it may change
depending on which agent is the health monitor leader.
For now, we just allow empty values on such state keys as
only one agent (the health monitor leader) will ask for these.

Change-Id: Id293e75e3aa47ec7a7335fade5a76f97854f0ce2
Ref: MINOR_CHANGE
Signed-off-by: Xavi León <xavi.leon@midokura.com>