Snapshot: use http header content-size
IOUtils.toByteArray cost is quite high when the length
of the data is unknown --i.e. writes to an output stream
in buffered chunks and once the whole stream is flushed,
all contents are copied back to a single byte array.

This patch takes the content size from the header
so copying from the input stream to the byte array
is done directly without buffering nor double copying.

For the content-size to be present, we must explicitly
disable content compression on the client. Otherwise,
the content-size is not written because data is compressed
on the fly, before the end of the stream is reached.

Change-Id: Ibbeba3717fbfa2cd5bef1742cbaaa0c92fd8fa4b
Ref: MINOR_CHANGE
Signed-off-by: Xavi León <xavi.leon@midokura.com>