Snapshot: send the content size before actual data
IOUtils.toByteArray cost is quite high when the length
of the data is unknown --i.e. writes to an output stream
in buffered chunks and once the whole stream is flushed,
all contents are copied back to a single byte array.

This patch sends the size of the stream before
the actual data so it directly allocates the whole
byte array and copies the data there directly.

Change-Id: Ibbeba3717fbfa2cd5bef1742cbaaa0c92fd8fa4b
Ref: MINOR_CHANGE
Signed-off-by: Xavi León <xavi.leon@midokura.com>