Snapshot: use http header content-size
IOUtils.toByteArray cost is quite high when the length
of the data is unknown --i.e. writes to an output stream
in buffered chunks and once the whole stream is flushed,
all contents are copied back to a single byte array.

This patch takes the content size from the header
so copying from the input stream to the byte array
is done directly without buffering nor double copying.

Change-Id: Ibbeba3717fbfa2cd5bef1742cbaaa0c92fd8fa4b
Ref: MINOR_CHANGE
Signed-off-by: Xavi León <xavi.leon@midokura.com>