Delay initialization of id field description
We cache the descriptor of the id field of the protobuf message
and the POJO message (backwards compatibility) to access
it directly and avoid using reflection every time. However,
this imposes a non-negligible initialization time. As the Agent
does not query the id never, just delay the initialization for
when it's used for the first time for the other Midonet
components.

Change-Id: Id0c6d47169fe85fbbb68c0d455b621d5eb9267c2
Ref: MI-2007
Signed-off-by: Xavi León <xavi.leon@midokura.com>