Stop bgp daemons on shutdown
This patch fixes the shutdown sequence of the
Routing Manager Actor so the actor system can
be shutted down gracefully.

In essence, we now wait for the routing handlers
to finish the shutdown sequence (stop bgp proces,
remove net namespace, etc.) before continuing with
the gracefull actor system shutdown, so we make
sure that the resources are properly cleaned up.
Otherwise, there's a chance that the actor
system stops all its processes before the
bgp resources are properly cleaned up.

Change-Id: I7f51493ac53abf883c79db76fd8ce381e608e709
Ref: MI-2602
Signed-off-by: Xavi León <xavi.leon@midokura.com>