Fix runtime config observable
We combine the latest items from two observables to
build the runtime configuration. However, the Agent
does not bundle the schemas in its jar, so we were
combining with an empty observable which never
emitted any value, and thus the combine observable
never returned any valued either. Changing the empty
observable to a just operator with an empty config
fixes the issue.

Change-Id: I6e860b6aa9cdf438f51330a9df8549d557166bcd
Ref: MINOR_CHANGE
Signed-off-by: Xavi León <xavi.leon@midokura.com>