Updates for Insights plugin
This makes several improvements to the Insights interface:

* The interface is now a Guava service, allowing it to have
a service lifecycle and start with the other agent service.

* The flowSimulation method receives the PacketContext instead
of individual fields. This addresses the issue where
the current fields were not sufficient to create a flow record,
requiring the plugin to install itself as a FlowRecorder using
reflection. Since a flow recorder would receive the PacketContext
anyway, this modifies the Insights interface accordingly.

* Fixes a bug, where the flowDelete was not called for
failures to remove a flow.

Ref: MINOR_CHANGE
Ref: https://midobugs.atlassian.net/browse/IC-900

Change-Id: I988ae36c931fb91b70e3491a4cd515a5a2956471
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>