Always check the state class paths on boot
State class paths are scoped by the namespace -- i.e.
each agent has its own directory tree that needs to
be created. When we moved the initialization of all
ZK paths to the cluster, we didn't check anymore that
state paths for the agents were created which made
state objects modifications to fail due to a non-existing
node.

This patch makes sure that the state paths are always
created regardless of the component (agent, cluster)
doing the initialization.

Change-Id: I371160be28f011ed86eaaca81d0ec989b7becad7
Ref: MI-2654
Signed-off-by: Xavi León <xavi.leon@midokura.com>