Fast reboot: allow failures on vtpm
While stopping the primary agent, the virtual to
physical mapper tries to set the active state
of the currently bound ports. However, at this
point of the fast reboot, the secondary agent
already has updated the value, changing the owner
of the state key. In this case, it is safe to
ignore the failed updated.

Change-Id: I288d82174a6f4537068076a60a2d69e1ab02d84d
Ref: MI-2007
Signed-off-by: Xavi León <xavi.leon@midokura.com>