Fast reboot: coordinate stop/start of new agent
This patch introduces a new mechaism to reboot the
agent that allows Midonet to provide close to
zero down-time during a reboot. The mechanism is
based on starting a new agent while the previous one
is still processing traffic. The shutdown of the
old agent and the startup of the new agent is
coordinated as to reduce to the minimum the
downtime by sending the SIGUSR1 signal to the
watchdog daemon.

To start the fast reboot mechanism, the operator
just needs to issue the command:

$ mm-reboot

Change-Id: Ic7dbbec022c61f52213feeb362f783709499ef25
Ref: MI-2007
Signed-off-by: Xavi León <xavi.leon@midokura.com>