Stop routing handler actor scheduled events
There are periodic events to retrieve the status of
the bgp daemon that we don't cancel when we stop the
routing handler. This patch saves the Cancellable
so we can stop these events when stopping the actor.

Change-Id: I03786b936582b86931273ee5daa5562eac84c1b9
Ref: MINOR_CHANGE
Signed-off-by: Xavi León <xavi.leon@midokura.com>
(cherry picked from commit 385dc9f04f02e205c2ddc155d9ed6082f9b8151f)