Fix concurrent modification in metrics
If a thread tries a traversal operation on a
histogram snapshot while it is being build, then
a concurrent modification exception may occur.

We have added a lock to protect operations that
require traversing the snapshot to prevent this
situation; the alternative of creating a new
immutable histogram as snapshot is less attractive
as it can pollute the memory and cause more GC
activations.

Ref: MI-2648

Change-Id: I084abf92bec3e213b2fc8f166e67de7d084ed02c
Signed-off-by: Ernest Artiaga <ernest@midokura.com>