Agent: fix NPE in packet workflow
When the simulation ended with an Error (e.g. timeout), the
packet context is returned to the packet context pool. Before
returning it to the pool, the packet context was being cleared,
effectively setting to null all its fields, including the
packet field, the culprit of the NPE.

Because the packet context is reset every time we take it from
the context pool, there's no need to do it when we give it back,
and we allow the packet to be recorded normally, avoiding the
NPE.

Change-Id: I189dcb7d4ee58d4c0603b872a19c3105a7a00fc2
Ref: MINOR_CHANGE
Signed-off-by: Xavi León <xavi.leon@midokura.com>
(cherry picked from commit 251a09c2a834294a28da72997a3cfb99ccb9e207)