BGP: graceful stop without routing handlers
We were waiting for the routing handlers to stop
to notify the shutdown process when the routing
manager is done. However, if there are no
routing handlers, no message is sent back so
it waits for the timeout, delaying the shutdown
sequence. This patch checks if we have
routing handlers to stop, and in this case, send
the finish message immediately.

Change-Id: I1e997bf22618419c9bc036db0b0af4f754185863
Ref: MINOR_CHANGE
Signed-off-by: Xavi León <xavi.leon@midokura.com>