Only write port active info if DP number is valid
If we received two active port messages with the same
datapath port number, then we assign the same ip address
to both of them. This means that this mapping is overwritten
in the InstanceInfoMap we use to map ip addresses to
instances. When the messages that the ports became inactive
arrives, the first one removes the overwritten entry,
and the second fails because it assumes there will be
an entry there, causing the agent to crash.

Ref: MI-2692

Change-Id: Ib9f9dbc7ab5832f7ea7957a7ecd5b8f110d401aa
Signed-off-by: Mateo Hermida <mateo@midokura.com>