Re-implement RebootBarriers with a fifo
The previous implementation was relying on the "main" agent starting
the coordination earlier than the "backup" agent.  While it's usually
the case, there's nothing to guarantee it.
Also, there's no point to use ZK for the coordination between agents
running on the same node.

Note: This implementation raises an exception on timeout.  The previous
implementation returns a boolean to indicate a timeout but none of callers
were checking it.

Note: This implementation is incompatible with the previous one.  That is,
a normal reboot is necessary between the versions.  I don't think it's
a problem given the experimental state of the feature.

Ref: MNA-1188
Ref: MNA-1198
Signed-off-by: YAMAMOTO Takashi <yamamoto@midokura.com>
Change-Id: I5627478a45435ab1a3ecae73293ef163dd89e8a7