wdog: Remove unnecessary timeout tweak for fast reboot
Because we launch another ProcMon instance for the "backup" agent
synchronously, the event loop of the ProcMon instance for the "main"
agent has no chance to run.  Thus, this tweak doesn't make any
effects. (Well, until the "backup" agent exits. But it isn't the
intention of this code, I guess.)

Ref: MINOR_CHANGE
Signed-off-by: YAMAMOTO Takashi <yamamoto@midokura.com>
Change-Id: I9dd6e707eb4487bf701550301145042e2c23ece0