wdog: Fix a spurious timeout
Don't busy-loop on a closed pipe.  Simply remove such a descriptor
from the select set.  It's safe as we know that the agent will never
re-open the pipe.

ProcMon.run assumes that select_loop.loop takes 1 seconds in many cases.
While usually it's true, it isn't at the very end of the agent shutdown
process.  After the agent (or JVM, or even the kernel) closed the wdog pipe
descriptor, if you're unlucky, the wdog process busy-loop on its side of
the pipe and ProcMon.run timeouts immediately.  (If you're lucky, the wdog
process gets a SIGCHLD soon enough and exits in an expected way.)

Ref: MNA-1206
Signed-off-by: YAMAMOTO Takashi <yamamoto@midokura.com>
Change-Id: Ibe28b492583afa108a00b72cc66b0af81b1b387c